solution/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-rest-framework==0.1.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
docopt==0.6.2
drf-writable-nested==0.7.1
future==1.0.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
paho-mqtt==1.6.1
pbr==6.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pykwalify==1.8.0
pytest==7.2.2
python-box==6.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.9.2
serializers==0.2.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stevedore==4.1.1
tavern==2.11.0
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
